Añadiendo errores en la funcion addpriv

+++ modules/m_authadd.php
@@ -73,6 +73,8 @@ class amodkey{
 		if(!isset($data->messageex[3])){$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, "\00305Error:\003 Faltan parámetros!!"); return 0;}
 		if($data->messageex[2]>9){$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, "\00305Error:\003 Error de usuario. Inserte otro usuario y presione enter. (No se pueden otorgar privilegios de nivel 10!!)"); return 0;}
 		$user = ORM::for_table('users')->where('username', strtolower($data->messageex[1]))->find_one(); 
+		if(!$user){$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, "\00305Error:\003 El usuario no existe!"); return 0;}
+		if(!is_numeric($data->messageex[2])){$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, "\00305Error:\003 Los privilegios deben ser un número entre el cero y el nueve."); return 0;}
 		$k = ORM::for_table('userpriv')->where('uid', $user->id)->where('sec',$data->messageex[3])->find_one();
 		if(method_exists($k, "delete")){$k->delete();}
 		$priv = ORM::for_table('userpriv')->create();
